pyproject.toml
[project]
name = "my-langgraph-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "ipykernel>=6.29.5",
    "langchain-google-genai>=2.1.4",
    "langchain-google-vertexai>=2.0.21",
    "langchain[anthropic]>=0.3.25",
    "langgraph>=0.4.1",
    "pgmpy>=1.0.0",
    "python-dotenv>=1.1.0",
    "requests",
]
